requirements.txt
html2text
pyyaml
requests
pdfkit